--- my_project/.github/cookiecutter.json
@@ -1,52 +1,0 @@
-{
-    "project_name": "my-project", 
-    "project_org": "DSACMS", 
-    "short_description": "A short description of the project.",
-    "long_description": "A longer description of the project.",
-    "subset_in_healthcare": "Policy, Operational",
-    "user_type": "Providers, Patients, Government",
-    "user_input": ["Yes", "No"],
-    "group": "CMS/OA/DSAC",
-    "repository_host": ["Github.com", "GitHub ENT", "GitHub Cloud", "GitLab.com", "GitLab ENT", "GitLab ENT CCSQ"],
-    "status": ["ideation", "development", "alpha", "beta", "release candidate", "production", "archival"],
-    "fisma_level": ["Low", "Moderate", "High"],
-    "license": ["CC0-1.0", "Apache-2.0", "MIT", "MPL-2.0", "GPL-2.0-only", "GPL-3.0-only", "GPL-3.0-or-later", "LGPL-2.1-only", "LGPL-3.0-only", "BSD-2-Clause", "BSD-3-Clause", "EPL-2.0", "Other"],
-    "vcs": ["git", "hg", "svn", "rcs", "bzr"],
-    "platforms": ["web", "windows", "mac", "linux", "ios", "android", "other"],
-    "categories": "healthcare",
-    "software_type":["standalone/mobile", "standalone/iot", "standalone/desktop", "standalone/web", "standalone/backend", "standalone/other", "addon", "library", "configurationFiles"],
-    "languages": "None",
-    "repo_type" : ["Package", "Website", "Standards", "Libraries", "Data", "Apps", "Tools", "APIs", "Docs"],
-    "usage_type" : ["openSource", "governmentWideReuse", "exemptByLaw", "exemptByNationalSecurity", "exemptByAgencySystem", "exemptByAgencyMission", "exemptByCIO", "exemptByPolicyDate"],
-    "maintenance": ["internal", "contract", "community", "none"],
-    "tags": "healthcare",
-    "contact_email": "opensource@cms.hhs.gov",
-    "contact_name": "CMS Open Source Program Office",
-    "localisation": ["true", "false"],
-    "__prompts__": {
-        "project_name": "Project name:",
-        "project_org": "GitHub organization:",
-        "short_description": "Provide a short description of the software. It should be a single line containing a single sentence. Maximum 150 characters are allowed.",
-        "long_description": "Provide longer description of the software, between 150 and 10000 chars. It is meant to provide an overview of the capabilities of the software for a potential user.",
-        "subset_in_healthcare": "Which subset of healthcare does the project belong to?",
-        "user_type": "Who are the intended users?",
-        "user_input": "Does the project accept user input? (e.g. allows user to query a database, allows login by users, upload files, etc.)",
-        "group": "Which group at CMS is the project part of?",
-        "repository_host": "Where is the repository hosted?",
-        "status": "What is the status of the project?",
-        "fisma_level": "What FISMA level is this project classified as? Learn more: https://security.cms.gov/learn/federal-information-security-modernization-act-fisma#perform-system-risk-categorization",
-        "license": "What license is the project under?",
-        "vcs": "What version control system is used?",
-        "platforms": "What platform does the software runs on?",
-        "categories": "What categories best describes the project? Separate items by commas. List of categories here: https://yml.publiccode.tools/categories-list.html?highlight=categories",
-        "software_type": "What type of software is the project?",
-        "languages": "What programming language(s) is the software written in? Separate items by commas.",
-        "repo_type": "What type of repository is this project?",
-        "usage_type": "What is the usage type for this project?",
-        "maintenance": "How is the software maintained?",
-        "tags": "Provide a list of tags to describe the software for search. Separate items by commas.",
-        "contact_name": "A point of contact is needed for the project. What is the name of the point of contact?",
-        "contact_email": "What is email address of the point of contact?",
-        "localisation": "Does the software support multiple spoken languages?"
-    }
-}
