--- my_project/.github/workflows/auto-changelog.yml
@@ -1,15 +1,0 @@
-name: Changelog
-on:
-  release:
-    types:
-      - created
-jobs:
-  changelog:
-    runs-on: ubuntu-latest
-    steps:
-      - name: "Auto Generate changelog"
-        uses: heinrichreimer/action-github-changelog-generator@v2.3
-        with:
-          
-          token: ${{ secrets.GITHUB_TOKEN }}
-
