--- my_project/CODEOWNERS.md
@@ -1,32 +1,0 @@
-# Code Owners
-<!-- TODO: Who are the points of contact in your project who are responsible/accountable for the project? This can often be an engineering or design manager or leader, who may or may not be the primary maintainers of the project. List them by GitHub Username-->
-
-- Git usernames of code owners; separated by commas.
-
-
-## Repo Domains
-
-<!--
-The Repo Domains section of your CODEOWNERS.md file helps manage code review responsibilities efficiently. Each domain represents a different aspect of the repository, such as documentation, frontend, backend, DevOps, testing, etc. In this section, list each domain and assign the appropriate GitHub usernames or teams responsible for that domain. This ensures that pull requests (PRs) are reviewed by the right experts, maintaining high code quality and relevance.
-
-For example:
-
-/docs/ @doc-team @johnsmith @janedoe
-
-/frontend/ @frontend-team @alice @bob
-
-/backend/ @backend-team @charlie @dana
-
-Furthermore, GitHub teams are a good feature for managing groups of contributors who need to be notified about specific domains within a repository. By creating and using GitHub teams, you can allow contributors to ping multiple relevant experts simultaneously.
-
-To set up GitHub teams:
-
-- Navigate to your organization's settings and select "Teams".
-- Create a new team for each domain, such as @frontend-team, @backend-team, or @doc-team.
-- Add the relevant members to each team. Ensure that the team includes all the individuals who should be notified about PRs in their domain.
-- When filling out the Repo Domains section in your CODEOWNERS.md file, use the team handles instead of or alongside individual usernames. This way, when a contributor opens a PR affecting a specific domain, they can simply tag the team, and every member of that team will be notified.
-
--->
-
-/docs/ {Git usernames of documentation owners}  
-/frontend/ {Git usernames of frontend owners}
