--- my_project/COMMUNITY_GUIDELINES.md
@@ -1,37 +1,0 @@
-# My Project Open Source Community Guidelines
-
-This document contains principles and guidelines for participating in the My Project open source community.
-
-## Principles
-
-These principles guide our data, product, and process decisions, architecture, and approach.
-
-- Open means transparent and participatory.
-- We take a modular and modern approach to software development.
-- We build open-source software and open-source process.
-- We value ease of implementation.
-- Fostering community includes building capacity and making our software and processes accessible to participants with diverse backgrounds and skillsets.
-- Data (and data science) is as important as software and process. We build open data sets where possible.
-- We strive for transparency for algorithms and places we might be introducing bias.
-
-## Community Guidelines
-
-All community members are expected to adhere to our [Code of Conduct](CODE_OF_CONDUCT.md).
-
-Information on contributing to this repository is available in our [Contributing file](CONTRIBUTING.md).
-
-When participating in My Project open source community conversations and spaces, we ask individuals to follow the following guidelines:
-
-- When joining a conversation for the first time, please introduce yourself by providing a brief intro that includes:
-  - your related organization (if applicable)
-  - your pronouns
-  - your superpower, and how you hope to use it for My Project
-- Embrace a culture of learning, and educate each other. We are all entering this conversation from different starting points and with different backgrounds. There are no dumb questions.
-- Take space and give space. We strive to create an equitable environment in which all are welcome and able to participate. We hope individuals feel comfortable voicing their opinions and providing contributions and will do our best to recognize and make space for individuals who may be struggling to find space here. Likewise, we expect individuals to recognize when they are taking up significant space and take a step back to allow room for others.
-<!-- TODO: Add if your repo has a community chat - Be present when joining synchronous conversations such as our community chat. Why be here if you're not going to _be here_? -->
-- Be respectful.
-- Default to positive. Assume others' contributions are legitimate and valuable and that they are made with good intention.
-
-## Acknowledgements
-
-This COMMUNITY_GUIDELINES.md was originally forked from the [United States Digital Service](https://usds.gov) [Justice40](https://thejustice40.com) open source [repository](https://github.com/usds/justice40-tool), and we would like to acknowledge and thank the community for their contributions.
