--- my_project/CONTRIBUTING.md
@@ -1,185 +1,0 @@
-<!--- # NOTE: Modify sections marked with `TODO` -->
-
-# How to Contribute
-
-<!-- Basic instructions about where to send patches, check out source code, and get development support.-->
-
-We're so thankful you're considering contributing to an [open source project of
-the U.S. government](https://code.gov/)! If you're unsure about anything, just
-ask -- or submit the issue or pull request anyway. The worst that can happen is
-you'll be politely asked to change something. We appreciate all friendly
-contributions.
-
-We encourage you to read this project's CONTRIBUTING policy (you are here), its
-[LICENSE](LICENSE.md), and its [README](README.md).
-
-## Getting Started
-
-<!--- TODO: If you have 'good-first-issue' or 'easy' labels for newcomers, mention them here.-->
-
-### Team Specific Guidelines
-
-<!-- TODO: This section helps contributors understand any team structure in the project (formal or informal.) Encouraged to point towards the MAINTAINERS.md file for further details.-->
-
-### Building dependencies
-
-<!--- TODO: This step is often skipped, so don't forget to include the steps needed to install on your platform. If you project can be multi-platform, this is an excellent place for first time contributors to send patches!-->
-
-### Building the Project
-
-<!--- TODO: Be sure to include build scripts and instructions, not just the source code itself! -->
-
-### Workflow and Branching
-
-<!--- TODO: Workflow Example
-We follow the [GitHub Flow Workflow](https://guides.github.com/introduction/flow/)
-
-1.  Fork the project
-2.  Check out the `main` branch
-3.  Create a feature branch
-4.  Write code and tests for your change
-5.  From your branch, make a pull request against `DSACMS/my-project/main`
-6.  Work with repo maintainers to get your change reviewed
-7.  Wait for your change to be pulled into `DSACMS/my-project/main`
-8.  Delete your feature branch
--->
-
-### Testing Conventions
-
-<!--- TODO: Discuss where tests can be found, how they are run, and what kind of tests/coverage strategy and goals the project has. -->
-
-### Coding Style and Linters
-
-<!--- TODO: HIGHLY ENCOURAGED. Specific tools will vary between different languages/frameworks (e.g. Black for python, eslint for JavaScript, etc...)
-
-1. Mention any style guides you adhere to (e.g. pep8, etc...)
-2. Mention any linters your project uses (e.g. flake8, jslint, etc...)
-3. Mention any naming conventions your project uses (e.g. Semantic Versioning, CamelCasing, etc...)
-4. Mention any other content guidelines the project adheres to (e.g. plainlanguage.gov, etc...)
-
--->
-
-### Writing Issues
-
-<!--- TODO: Example Issue Guides
-
-When creating an issue please try to adhere to the following format:
-
-    module-name: One line summary of the issue (less than 72 characters)
-
-    ### Expected behavior
-
-    As concisely as possible, describe the expected behavior.
-
-    ### Actual behavior
-
-    As concisely as possible, describe the observed behavior.
-
-    ### Steps to reproduce the behavior
-
-    List all relevant steps to reproduce the observed behavior.
-
-    see our .github/ISSUE_TEMPLATE.md for more examples.
--->
-
-### Writing Pull Requests
-
-<!-- TODO: Make a brief statement about where to file pull/merge requests, and conventions for doing so. Link to PULL_REQUEST_TEMPLATE.md file.
-
-Comments should be formatted to a width no greater than 80 columns.
-
-Files should be exempt of trailing spaces.
-
-We adhere to a specific format for commit messages. Please write your commit
-messages along these guidelines. Please keep the line width no greater than 80
-columns (You can use `fmt -n -p -w 80` to accomplish this).
-
-    module-name: One line description of your change (less than 72 characters)
-
-    Problem
-
-    Explain the context and why you're making that change.  What is the problem
-    you're trying to solve? In some cases there is not a problem and this can be
-    thought of being the motivation for your change.
-
-    Solution
-
-    Describe the modifications you've done.
-
-    Result
-
-    What will change as a result of your pull request? Note that sometimes this
-    section is unnecessary because it is self-explanatory based on the solution.
-
-Some important notes regarding the summary line:
-
-* Describe what was done; not the result
-* Use the active voice
-* Use the present tense
-* Capitalize properly
-* Do not end in a period — this is a title/subject
-* Prefix the subject with its scope
-
-    see our .github/PULL_REQUEST_TEMPLATE.md for more examples.
--->
-
-## Reviewing Pull Requests
-
-<!--- TODO: Make a brief statement about how pull-requests are reviewed, and who is doing the reviewing. Linking to MAINTAINERS.md can help.
-
-Code Review Example
-
-The repository on GitHub is kept in sync with an internal repository at
-github.cms.gov. For the most part this process should be transparent to the
-project users, but it does have some implications for how pull requests are
-merged into the codebase.
-
-When you submit a pull request on GitHub, it will be reviewed by the project
-community (both inside and outside of github.cms.gov), and once the changes are
-approved, your commits will be brought into github.cms.gov's internal system for
-additional testing. Once the changes are merged internally, they will be pushed
-back to GitHub with the next sync.
-
-This process means that the pull request will not be merged in the usual way.
-Instead a member of the project team will post a message in the pull request
-thread when your changes have made their way back to GitHub, and the pull
-request will be closed.
-
-The changes in the pull request will be collapsed into a single commit, but the
-authorship metadata will be preserved.
-
--->
-
-
-## Shipping Releases
-
-<!-- TODO: What cadence does your project ship new releases? (e.g. one-time, ad-hoc, periodically, upon merge of new patches) Who does so? -->
-
-
-## Documentation
-
-<!-- TODO: Documentation Example
-
-We also welcome improvements to the project documentation or to the existing
-docs. Please file an [issue](https://github.com/DSACMS/my-project/issues).
--->
-
-## Policies
-
-### Open Source Policy
-
-We adhere to the [CMS Open Source
-Policy](https://github.com/CMSGov/cms-open-source-policy). If you have any
-questions, just [shoot us an email](mailto:opensource@cms.hhs.gov).
-
-### Security and Responsible Disclosure Policy
-
-*Submit a vulnerability:* Vulnerability reports can be submitted through [Bugcrowd](https://bugcrowd.com/cms-vdp). Reports may be submitted anonymously. If you share contact information, we will acknowledge receipt of your report within 3 business days.
-
-For more information about our Security, Vulnerability, and Responsible Disclosure Policies, see [SECURITY.md](SECURITY.md).
-
-## Public domain
-
-This project is in the public domain within the United States, and copyright and related rights in the work worldwide are waived through the [CC0 1.0 Universal public domain dedication](https://creativecommons.org/publicdomain/zero/1.0/) as indicated in [LICENSE](LICENSE).
-
-All contributions to this project will be released under the CC0 dedication. By submitting a pull request or issue, you are agreeing to comply with this waiver of copyright interest.
