+++ tier4/{{cookiecutter.project_slug}}/.github/codejson/{{cookiecutter.project_name}}/code.json
@@ -13,7 +13,9 @@
         "URL": "LICENSE",
         "name": "{{ cookiecutter.license }}"
       }
-    ]
+    ],
+    "usageType": "{{ cookiecutter.usage_type }}",
+    "exemptionText": ""
   },
   "organization": "Centers for Medicare & Medicaid Services",
   "repositoryURL": "https://github.com/{{ cookiecutter.project_org }}/{{ cookiecutter.project_name }}",
@@ -33,8 +35,6 @@
   "localisation": "{{ cookiecutter.localisation }}",
   "repoType": "{{ cookiecutter.repo_type }}",
   "userInput": "{{ cookiecutter.user_input }}",
-  "usageType": "{{ cookiecutter.usage_type }}",
-  "exemptionText": "",
   "fismaLevel": "{{ cookiecutter.fisma_level }}",
   "group": "{{ cookiecutter.group }}",
   "subsetInHealthcare": "{{ cookiecutter.subset_in_healthcare }}",
